Update tera.json
Is it possible to apply the tera vscode extension to `___.html.tera` files with this change? 

I'm using (rocket)[https://rocket.rs] and for clarity I'm just using the .html.tera extension.  Right now the below files are just appearing as raw text files in vscode.

Apologies if this has been covered already, I didn't see a PR or Issue with `html.tera`.

```
./templates
├── base.html.tera
├── login.html.tera
├── navbar.html.tera
└── new_sheet.html.tera
```